Replace "−" with "-" in Assigment 3 tutorial text
When viewing this file in VSCode, I saw the following warning highlighting lines `15` and `29`

```
The character U+2212 "−" could be confused with the character U+002d "-", which is more common in source code.
```

It seems to me this is just a small oversight, as "-" is used elsewhere in the commented block to denote a negative value, but I'll freely admit ignorance as to any nuance here in using one vs. the other.