Add the find-zig-release-branch.sh tool
The tool is used to find the latest commit, from the HEAD in date order,
that is compatible with the specified Zig compiler.

The commit found can be used as a base for a branch that can be used by
users that can only install a released version of Zig, e.g. v0.10.1.